requirements.txt
pdfplumber>=0.7.4
python-docx>=0.8.11
pillow>=9.0.0
pytesseract>=0.3.10
spacy>=3.5.0
networkx>=3.0
sqlalchemy>=1.4
tqdm>=4.0
regex
langdetect
fastapi
uvicorn